Update
Unfortunately the FTP program I’m using to update the site seems to be
incompatible with Github so I have to copy all of the files over every
once in a while…

+++ view/event_creator_footer.php
@@ -11,7 +11,7 @@
 <div class="footer creator_footer">
 	<div class="footer_left">
 		<a class="button light_button" href=".">Cancel</a><br />
-		<a class="button light_button" href=".?a=delete&e=<?php echo $event_id; ?>">Delete Event</a>
+		<a class="button light_button" href=".?a=delete&e=<?php echo $event_id; ?>" onclick="return confirm('Are you sure you wish to delete this event? (This cannot be undone)');">Delete Event</a>
     </div>
     <div class="footer_right">
     	<input class="button light_button" type="submit" name="action" value="Reminders"/><br />
